refactor(lsp): extract shared whole-document bridged fan-out
documentLink, foldingRange, and codeLens each carried an identical
~150-line region fan-out (resolve regions, preferred dispatch per
region, concatenate, cancel handling). A third copy crossed the
duplication threshold: future whole-document methods and fixes to the
fan-out (e.g. cancel semantics) should land in one place. Handlers now
supply only the method name and the downstream send call.
