+++ src/lsp/lsp_impl/text_document.rs
@@ -33,5 +33,6 @@ mod selection_range;
 mod semantic_tokens;
 mod signature_help;
 mod type_definition;
+mod whole_document;
 
 // Re-export the methods (they are implemented as impl blocks on Kakehashi)
